+++ simple_history/models.py
@@ -567,7 +567,7 @@ class HistoricalRecords:
             """
             Get the next history record for the instance. `None` if last.
             """
-            history = utils.get_history_manager_from_history(self)
+            history = utils.get_historical_records_of_instance(self)
             return (
                 history.filter(history_date__gt=self.history_date)
                 .order_by("history_date")
@@ -578,7 +578,7 @@ class HistoricalRecords:
             """
             Get the previous history record for the instance. `None` if first.
             """
-            history = utils.get_history_manager_from_history(self)
+            history = utils.get_historical_records_of_instance(self)
             return (
                 history.filter(history_date__lt=self.history_date)
                 .order_by("history_date")